requirements.txt
aiofiles
aiogram~=3.18.0
aiohttp
aiosignal
alembic~=1.15.1
async-timeout
attrs
certifi
charset-normalizer
frozenlist
greenlet
idna
magic-filter
Mako
MarkupSafe
multidict
psycopg
psycopg-binary
psycopg-pool
pydantic~=2.10.6
python-dotenv
SQLAlchemy~=2.0.38
typing_extensions
yarl
pydantic-settings~=2.8.1
geopy

Django~=5.1.6
pillow~=11.1.0